DOC: Adding documentation.
